Proper history-substring-search and other fixes.
Finally proper history-substring-search which also works with vi keys.
Furthermore I removed the number.vim plugin as it is responsible for
more confusion than it solves problems.
Also added "set scrolloff=5" to the vimrc to keep more lines in view.
General cleanup and some added functionality in the vim and zsh config
files.